Move ssh-agent startup to .bashrc.
Starting ssh-agent in .xinitrc was troublesome if the agent ever needed
restarting as all terminals inherited the incorrect PID and socket. This way the
agent is started and the key loaded the first time you open a terminal after
boot. If you need to kill the agent for any reason that is now possible too.